salvar

+++ ecommerce/src/main/java/br/edu/unfio/ecommerce/entidades/ItemPedido.java
@@ -20,7 +20,7 @@ public class ItemPedido {
  
     private Integer id;
     private Integer quantidade;
-    private BigDecimal valorUnitario;
+    private Integer valorUnitario;
 
     @ManyToOne
     private Pedido pedido;
